second

+++ lib/models/movies.dart
@@ -18,13 +18,14 @@ class Movies {
 
   factory Movies.fromJson(Map<String, dynamic> json) {
     return Movies(
-        title: json["title"],
-        backDropPath: json["backdrop_path"],
-        originalTitle: json["original_title"],
-        overView: json["overview"],
-        posterPath: json["poster_path"],
-        releaseDate: json["release_date"],
-        voteAverage: json["vote_average"]);
+      title: json["title"],
+      backDropPath: json["backdrop_path"],
+      originalTitle: json["original_title"],
+      overView: json["overview"],
+      posterPath: json["poster_path"],
+      releaseDate: json["release_date"],
+      voteAverage: json["vote_average"].toDouble(),
+    );
   }
 }
